+++ packages/studio-base/src/players/IterablePlayer/CoSceneIterablePlayer.ts
@@ -37,6 +37,7 @@ import {
   PlayerStateActiveData,
 } from "@foxglove/studio-base/players/types";
 import { RosDatatypes } from "@foxglove/studio-base/types/RosDatatypes";
+import { getPlaybackQualityLevelByLocalStorage } from "@foxglove/studio-base/util/coscene";
 import delay from "@foxglove/studio-base/util/delay";
 
 import { BlockLoader } from "./BlockLoader";
@@ -551,25 +552,8 @@ export class CoSceneIterablePlayer implements Player {
     // set the playIterator to the seek time
     await this._bufferedSource.stopProducer();
 
-    const localPlaybackQualityLevel = localStorage.getItem("playbackQualityLevel");
-    let playbackQualityLevel: "ORIGINAL" | "HIGH" | "MID" | "LOW" = "ORIGINAL";
-
-    switch (localPlaybackQualityLevel) {
-      case "ORIGINAL":
-        playbackQualityLevel = "ORIGINAL";
-        break;
-      case "HIGH":
-        playbackQualityLevel = "HIGH";
-        break;
-      case "MID":
-        playbackQualityLevel = "MID";
-        break;
-      case "LOW":
-        playbackQualityLevel = "LOW";
-        break;
-      default:
-        playbackQualityLevel = "ORIGINAL";
-    }
+    const playbackQualityLevel: "ORIGINAL" | "HIGH" | "MID" | "LOW" =
+      getPlaybackQualityLevelByLocalStorage();
 
     log.debug("Initializing forward iterator from", next);
     this._playbackIterator = this._bufferedSource.messageIterator({
@@ -611,25 +595,8 @@ export class CoSceneIterablePlayer implements Player {
       throw new Error("Invariant. playbackIterator was already set");
     }
 
-    const localPlaybackQualityLevel = localStorage.getItem("playbackQualityLevel");
-    let playbackQualityLevel: "ORIGINAL" | "HIGH" | "MID" | "LOW" = "ORIGINAL";
-
-    switch (localPlaybackQualityLevel) {
-      case "ORIGINAL":
-        playbackQualityLevel = "ORIGINAL";
-        break;
-      case "HIGH":
-        playbackQualityLevel = "HIGH";
-        break;
-      case "MID":
-        playbackQualityLevel = "MID";
-        break;
-      case "LOW":
-        playbackQualityLevel = "LOW";
-        break;
-      default:
-        playbackQualityLevel = "ORIGINAL";
-    }
+    const playbackQualityLevel: "ORIGINAL" | "HIGH" | "MID" | "LOW" =
+      getPlaybackQualityLevelByLocalStorage();
 
     log.debug("Initializing forward iterator from", this._start);
     this._playbackIterator = this._bufferedSource.messageIterator({
@@ -721,25 +688,8 @@ export class CoSceneIterablePlayer implements Player {
 
     try {
       this._abort = new AbortController();
-      const localPlaybackQualityLevel = localStorage.getItem("playbackQualityLevel");
-      let playbackQualityLevel: "ORIGINAL" | "HIGH" | "MID" | "LOW" = "ORIGINAL";
-
-      switch (localPlaybackQualityLevel) {
-        case "ORIGINAL":
-          playbackQualityLevel = "ORIGINAL";
-          break;
-        case "HIGH":
-          playbackQualityLevel = "HIGH";
-          break;
-        case "MID":
-          playbackQualityLevel = "MID";
-          break;
-        case "LOW":
-          playbackQualityLevel = "LOW";
-          break;
-        default:
-          playbackQualityLevel = "ORIGINAL";
-      }
+      const playbackQualityLevel: "ORIGINAL" | "HIGH" | "MID" | "LOW" =
+        getPlaybackQualityLevelByLocalStorage();
       const messages = await this._bufferedSource.getBackfillMessages({
         topics,
         time: targetTime,
